// src/common/types.ts
// Copyright 2011 Mark Cavage <mcavage@gmail.com> All rights reserved.

/**
 * See http://www.oss.com/asn1/resources/reference/asn1-reference-card.html
 */
export enum ASN1 {
  EOC = 0,
  Boolean = 1,
  Integer = 2,
  BitString = 3,
  OctetString = 4,
  Null = 5,
  OID = 6,
  ObjectDescriptor = 7,
  External = 8,
  Real = 9, // float
  Enumeration = 10,
  PDV = 11,
  Utf8String = 12,
  RelativeOID = 13,
  Sequence = 16,
  Set = 17,
  NumericString = 18,
  PrintableString = 19,
  T61String = 20,
  VideotexString = 21,
  IA5String = 22,
  UTCTime = 23,
  GeneralizedTime = 24,
  GraphicString = 25,
  VisibleString = 26,
  GeneralString = 28,
  UniversalString = 29,
  CharacterString = 30,
  BMPString = 31,
  Constructor = 32,
  Context = 128
};